a few comments
