+++ systemtest/src/main/java/io/strimzi/systemtest/performance/utils/TopicOperatorPerformanceUtils.java
@@ -13,6 +13,7 @@ import io.strimzi.systemtest.resources.ResourceManager;
 import io.strimzi.systemtest.resources.crd.KafkaTopicResource;
 import io.strimzi.systemtest.storage.TestStorage;
 import io.strimzi.systemtest.utils.kafkaUtils.KafkaTopicScalabilityUtils;
+import io.strimzi.systemtest.utils.kafkaUtils.KafkaTopicUtils;
 import io.strimzi.test.TestUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -138,74 +139,8 @@ public class TopicOperatorPerformanceUtils {
     private static void performDeletionWithWait(int start, int end, ExtensionContext currentContext, TestStorage testStorage) {
         ResourceManager.setTestContext(currentContext);
         LOGGER.info("Deleting Kafka topics from index {} to {}", start, end);
-        deleteKafkaTopicsInRange(testStorage.getNamespaceName(), testStorage.getTopicName(), start, end);
-        waitForTopicWithPrefixDeletion(testStorage.getNamespaceName(), testStorage.getTopicName(), start, end);
-    }
-
-    /**
-     * Deletes Kafka topics within a specified range based on their names.
-     * Assumes topic names are structured to include a numerical suffix indicating their batch.
-     *
-     * @param namespace     the Kubernetes namespace
-     * @param prefix        the common prefix of the Kafka topic names
-     * @param start         the starting index of topics to delete
-     * @param end           the ending index of topics to delete
-     */
-    public static void deleteKafkaTopicsInRange(String namespace, String prefix, int start, int end) {
-        List<KafkaTopic> topicsToDelete = KafkaTopicResource.kafkaTopicClient().inNamespace(namespace).list().getItems()
-            .stream()
-            .filter(topic -> {
-                String name = topic.getMetadata().getName();
-                if (!name.startsWith(prefix)) return false;
-                // Extract index from the topic name
-                String indexPart = name.substring(prefix.length());
-                try {
-                    int index = Integer.parseInt(indexPart);
-                    return index >= start && index < end;
-                } catch (NumberFormatException e) {
-                    LOGGER.error("Failed to parse index from Kafka topic name: {}", name, e);
-                    return false;
-                }
-            }).toList();
-
-        topicsToDelete.forEach(topic ->
-            cmdKubeClient().namespace(namespace).deleteByName(KafkaTopic.RESOURCE_SINGULAR, topic.getMetadata().getName())
-        );
-
-        LOGGER.info("Deleted Kafka topics from {} to {} in namespace {} with prefix {}", start, end, namespace, prefix);
-    }
-
-    public static List<KafkaTopic> getAllKafkaTopicsWithPrefix(String namespace, String prefix, int start, int end) {
-        return KafkaTopicResource.kafkaTopicClient().inNamespace(namespace).list().getItems()
-            .stream()
-            .filter(topic -> {
-                String name = topic.getMetadata().getName();
-                if (!name.startsWith(prefix)) return false;
-                String indexPart = name.substring(prefix.length());
-                try {
-                    int index = Integer.parseInt(indexPart);
-                    return index >= start && index < end;
-                } catch (NumberFormatException e) {
-                    LOGGER.error("Failed to parse index from Kafka topic name: {}", name, e);
-                    return false;
-                }
-            })
-            .collect(Collectors.toList());
-    }
-
-    public static void waitForTopicWithPrefixDeletion(String namespaceName, String topicPrefix, int start, int end) {
-        TestUtils.waitFor("deletion of all topics with prefix: " + topicPrefix + " from " + start + " to " + end,
-            TestConstants.GLOBAL_POLL_INTERVAL, TestConstants.GLOBAL_TIMEOUT,
-            () -> {
-                try {
-                    final int numberOfTopicsToDelete = getAllKafkaTopicsWithPrefix(namespaceName, topicPrefix, start, end).size();
-                    LOGGER.info("Remaining KafkaTopic's to delete from {} to {}: {} !", start, end, numberOfTopicsToDelete);
-                    return numberOfTopicsToDelete == 0;
-                } catch (Exception e) {
-                    LOGGER.error("Error while checking for topic deletion: ", e);
-                    return e.getMessage().contains("Not Found") || e.getMessage().contains("the server doesn't have a resource type");
-                }
-            });
+        KafkaTopicUtils.deleteKafkaTopicsInRange(testStorage.getNamespaceName(), testStorage.getTopicName(), start, end);
+        KafkaTopicUtils.waitForTopicWithPrefixDeletion(testStorage.getNamespaceName(), testStorage.getTopicName(), start, end);
     }
 
     /**
